refactor(e2e): rename baseline _fixtures package to fixtures
Drop the leading underscore; it's a normal package, not a private one.
Update conftest imports and doc comments to match.
