+++ tests/e2e/baseline/fixtures/capture.py
