+++ tests/e2e/baseline/fixtures/platform.py
